+++ modules/UData/HttpResponse/HttpResponse.hpp
@@ -3,10 +3,10 @@
 
 # include "../../config/ConfParser.hpp"
 # include "../HttpRequest/HttpRequest.hpp"
-# include "AutoIndex.hpp"
-# include "StatusMsgStore.hpp"
-# include "MimeStore.hpp"
-# include "../../Event/Session.hpp"
+# include "./Features/AutoIndex.hpp"
+# include "./Store/StatusMsgStore.hpp"
+# include "./Store/MimeStore.hpp"
+# include "./Features/Session.hpp"
 
 # include <iostream>
 # include <fstream>
